selinux: context switch to container_t
Switch the selinux context of the next spawned thread/process to
`container_t`.

The netns.Do{...} will be executed in a separate thread. As such,
we need the request the selinux for the next operations within that
block, by using `selinux.SetExecLabel(...)`.

A different alternative would have been to update the `current`
label, via `selinux.SetTaskLabel`, but unfortunately, that
operation requires dynamictransition permissions, which, AFAIK, we
are not interested in adding. As such, the former alternative was
chosen.

Furthermore, setting the selinux context (and being able to use it)
must be performed within the k8s node mount ns, since otherwise
AVCs are thrown while accessing the pfifo_queue and /dev/null.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>